# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Credit Repair Pro API",
    "description": "Complete API for Credit Repair SaaS — client management, credit tracking, dispute generation, payments, compliance (CROA/FCRA/GLBA), and AI-powered analysis.",
    "version": "1.0.0",
    "contact": {
      "name": "TriExpert Services",
      "url": "https://triexpertservice.com"
    }
  },
  "servers": [
    { "url": "https://triexpertservice.com/api", "description": "Production" },
    { "url": "http://localhost:5000/api", "description": "Development" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Local JWT or Auth0 token"
      }
    },
    "schemas": {
...
      "CreditItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "clientId": { "type": "string", "format": "uuid" },
          "itemType": { "type": "string", "enum": ["late_payment", "collection", "charge_off", "bankruptcy", "repossession", "foreclosure", "inquiry", "other"] },
          "creditorName": { "type": "string" },
          "accountNumber": { "type": "string" },
          "bureau": { "type": "string" },
          "balance": { "type": "number" },
          "status": { "type": "string", "enum": ["identified", "disputing", "in_progress", "resolved", "deleted"] }
        }
      },
...
    }
  },
  "paths": {
...
    "/credit-items": {
      "get": {
        "tags": ["Credit Items"], "summary": "Get current user's credit items", "operationId": "getMyCreditItems",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": { "200": { "description": "Paginated credit items" } }
      },
      "post": {
        "tags": ["Credit Items"], "summary": "Add new credit item", "operationId": "addCreditItem",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreditItem" } } } },
        "responses": { "201": { "description": "Credit item created" } }
      }
    },
    "/credit-items/client/{clientId}": {
      "get": {
        "tags": ["Credit Items"], "summary": "Get credit items for specific client", "operationId": "getClientCreditItems",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "clientId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Credit items for client" } }
      }
    },
    "/credit-items/{id}/status": {
      "put": {
        "tags": ["Credit Items"], "summary": "Update credit item status", "operationId": "updateCreditItemStatus",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "string" } }, "required": ["status"] } } } },
        "responses": { "200": { "description": "Status updated" } }
      }
    },
    "/credit-items/{id}": {
      "delete": {
        "tags": ["Credit Items"], "summary": "Delete credit item", "operationId": "deleteCreditItem",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Credit item deleted" } }
      }
    },
...
    "/credit-reports/items/{clientId}": { "get": { "tags": ["Credit Reports"], "summary": "Get all credit items", "operationId": "getReportItems", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "clientId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Credit items" } } } },
...
    "/credit-reports/items/{itemId}": { "put": { "tags": ["Credit Reports"], "summary": "Update a credit item", "operationId": "updateReportItem", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "itemId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Item updated" } } }, "delete": { "tags": ["Credit Reports"], "summary": "Delete credit item + disputes", "operationId": "deleteReportItem", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "itemId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Item deleted" } } } },
...
  }
}